Grid System, Building and Destroying, Ghost Highlight
Made an infinite grid system using a dictionary so that only the grid spaces that are occupied are saved, it should be able to save and rebuild the grid spaces as well.

Building and Destroying works, I made a placeholder Bee house for this process. Ghost preview of bee house also available.