Complete Inventory Overhaul + Item Icon Generation
Implemented an entire inventory overhaul, removing the entirety of the old system in favor of a new, simpler system.

New scene for Icon Generation, which, after putting objects into the scene, automatically generates icons for the inventory system.

TODO:
Remove Ghost Preview of object when there is none left in inventory.